refactor: Rename getSuccessMessage into a more generic name
