+++ packages/cozy-sharing/src/helpers/share.js
